% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-tests.R
\name{test_column_between}
\alias{test_column_between}
\title{Test a column is within an expected range}
\usage{
test_column_between(df, col, min_expected, max_expected)
}
\arguments{
\item{df}{A dataframe}

\item{col}{The unquoted column name of the column you want to test}

\item{min_expected}{The expected minimum value}

\item{max_expected}{The expected maximum value}
}
\value{
The dataframe passed to the function.
}
\description{
Performs a test that no value in a specified column is
greater or smaller than an expected minimum and maximum
value.
}
\details{
This function is also good for verifying that rescaling operations
have worked as expected.
}
\examples{
# Basic usage
test_column_between(mtcars, disp, 50, 500)

# Test a scaling operation
mtcars \%>\%
mutate(disp_scaled = (disp - min(disp)) / (max(disp) - min(disp))) \%>\%
test_column_between(disp_scaled, 0, 1)

}
