% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-tests.R
\name{test_column_max}
\alias{test_column_max}
\title{Test maximum value of a column}
\usage{
test_column_max(df, col, max_expected)
}
\arguments{
\item{df}{A dataframe}

\item{col}{The unquoted column name of the column you want to test}

\item{max_expected}{The expected maximum value}
}
\value{
The dataframe passed to the function.
}
\description{
Performs a test that no value in a specific column is larger
than an expected value.
}
\details{
Note that this test only verifies that all values in the column are
less than or equal to the specified value. It does \strong{not} test that the
maximum value of the column is equal to a certain value.
}
\examples{
# Basic usage
test_column_max(mtcars, disp, 500)

}
