% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-tests.R
\name{test_column_n_unique_values}
\alias{test_column_n_unique_values}
\title{Test unique values in a column}
\usage{
test_column_n_unique_values(df, col, n_expected)
}
\arguments{
\item{df}{A dataframe}

\item{col}{The unquoted column name of the column you want to test}

\item{n_expected}{The expected number of unique values}
}
\value{
The dataframe passed to the function.
}
\description{
Performs a test that a specific column has an expected number
of unique values.
}
\details{
This function is good to utilize after performing joining operations
to validate that no duplicates or missing values have been created. Note
that this function does not count how many times each unique value
appears - just the overall number of distinct items.
}
\examples{
# Basic usage
test_column_n_unique_values(mtcars, cyl, 3)

}
