% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-tests.R
\name{test_column_sum}
\alias{test_column_sum}
\title{Test the sum of a column}
\usage{
test_column_sum(df, col, sum)
}
\arguments{
\item{df}{A dataframe}

\item{col}{The unquoted column name of the column you want to test}

\item{sum}{The expected sum of the column}
}
\value{
The dataframe passed to the function.
}
\description{
Performs a test that a specific column has an expected sum.
}
\details{
This function is good to utilize after performing joining operations
to validate that no duplicates or missing values have been created.
}
\examples{
# Basic usage
test_column_sum(mtcars, wt, 102.953)

}
