% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-tests.R
\name{test_complete}
\alias{test_complete}
\title{Test that there are no missing values in a dataframe}
\usage{
test_complete(df, ...)
}
\arguments{
\item{df}{A dataframe}

\item{...}{Optional selection of specific columns to include in the
duplicate check.}
}
\value{
The dataframe passed to the function.
}
\description{
Performs a test that a dataframe has no duplicate rows.
Column names can be passed in (like \code{\link[dplyr]{select}})
if you wish to restrict the duplicate search to just these columns.
}
\details{
When selecting a subset of columns be aware that duplicates will
still be looked for within the entire row - not for each column
individually.
}
\examples{
# Basic usage - passes
test_no_duplicates(mtcars)

# Select specific columns - fails
\dontrun{
test_no_duplicates(mtcars, cyl)
}

# Can also be used in magrittr pipe
library(dplyr)
mtcars \%>\% test_no_duplicates() \%>\% select(mpg)

}
