% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-tests.R
\name{test_equal_rows}
\alias{test_equal_rows}
\title{Test that two dataframes have the same number of rows}
\usage{
test_equal_rows(df, df2)
}
\arguments{
\item{df}{A dataframe}

\item{df2}{Another dataframe}
}
\value{
df
}
\description{
Performs a test that the dataframe has the number of rows as
another dataframe (even if the contents are different).
}
\examples{
# Basic usage
test_equal_rows(mtcars, mtcars)
test_equal_rows(mtcars, iris)

# Can also be used in magrittr pipe
mtcars \%>\%
mutate(mpg = 50) \%>\%
test_equal_rows(mtcars)

}
