% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-tests.R
\name{test_n_per_group}
\alias{test_n_per_group}
\title{Test that there n rows per group}
\usage{
test_n_per_group(df, n, ...)
}
\arguments{
\item{df}{A dataframe}

\item{n}{Expected number of rows}

\item{...}{Grouping variables}
}
\value{
The dataframe passed to the function.
}
\description{
Performs a test that a dataframe has no duplicate rows.
Column names can be passed in (like \code{\link[dplyr]{select}})
if you wish to restrict the duplicate search to just these columns.
}
\details{
When selecting a subset of columns be aware that duplicates will
still be looked for within the entire row - not for each column
individually.
}
\examples{
# Basic usage - passes
test_no_duplicates(mtcars)

# Select specific columns - fails
\dontrun{
test_no_duplicates(mtcars, cyl)
}

# Can also be used in magrittr pipe
library(dplyr)
mtcars \%>\% test_no_duplicates() \%>\% select(mpg)

}
