% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-tests.R
\name{test_n_rows}
\alias{test_n_rows}
\title{Test the number rows in a dataframe}
\usage{
test_n_rows(df, n)
}
\arguments{
\item{df}{A dataframe}

\item{n}{Expected number of rows}
}
\value{
The dataframe passed to the function.
}
\description{
Performs a test that the dataframe has the number of rows you expect.
}
\examples{
# Basic usage
test_n_rows(mtcars, 32)

# Can also be used in magrittr pipe
mtcars \%>\% test_n_rows(32) \%>\% select(mpg)

}
