% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-tests.R
\name{test_no_duplicates}
\alias{test_no_duplicates}
\title{Test that there are no duplicates in a dataframe}
\usage{
test_no_duplicates(df, ...)
}
\arguments{
\item{df}{A dataframe}

\item{...}{Optional selection of specific columns to include in the
duplicate check.}
}
\value{
The dataframe passed to the function.
}
\description{
Performs a test that a dataframe has
}
\examples{
# Basic usage
test_n_rows(mtcars, 32)

# Can also be used in magrittr pipe
mtcars \%>\% test_n_rows(32) \%>\% select(mpg)

}
