% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df-tests.R
\name{test_same_as_other_df}
\alias{test_same_as_other_df}
\title{Test that two dataframes are equal}
\usage{
test_same_as_other_df(df, df2)
}
\arguments{
\item{df}{A dataframe}

\item{df2}{Another dataframe}
}
\value{
df
}
\description{
Performs a test that the dataframe is equal to another dataframe.
}
\examples{
# Basic usage
test_same_as_other_df(mtcars, mtcars)
test_same_as_other_df(mtcars, iris)

# Can also be used in magrittr pipe
mtcars \%>\%
mutate(mpg = 50) \%>\%
test_same_as_other_df(mtcars)

}
