% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/col-tests.R
\name{test_n_per_group}
\alias{test_n_per_group}
\title{Test that there n rows per group}
\usage{
test_n_per_group(df, n, ...)
}
\arguments{
\item{df}{A dataframe}

\item{n}{Expected number of rows}

\item{...}{Grouping variables}
}
\value{
The dataframe passed to the function.
}
\description{
Performs a test that every grouped level of a dataframe has exactly
n rows.
}
\details{
You can use multiple grouping variables in this function.
}
\examples{
# Basic usage - passes
test_n_per_group(iris, 50, Species)

}
